change data type from `uint64_t` -> `size_t`
always use hwthread id no in debug statements

+++ include/bitmask.h
@@ -11,7 +11,7 @@
 #include <stdint.h>
 
 
-static inline int is_multipleof_pow2(const uint64_t in, const uint64_t nbits)
+static inline int is_multipleof_pow2(const size_t in, const size_t nbits)
 {
     if (nbits == 0 || (nbits & (nbits - 1)) != 0)
     {
@@ -20,7 +20,7 @@ static inline int is_multipleof_pow2(const uint64_t in, const uint64_t nbits)
     return ((in & (nbits - 1)) == 0);
 }
 
-static inline int is_multipleof_nbits(const uint64_t in, const uint64_t nbits)
+static inline int is_multipleof_nbits(const size_t in, const size_t nbits)
 {
     if (nbits == 0)
     {
@@ -29,13 +29,13 @@ static inline int is_multipleof_nbits(const uint64_t in, const uint64_t nbits)
     return (in % nbits == 0);
 }
 
-static inline int roundup_nbits(uint64_t* out, const uint64_t in, const uint64_t nbits)
+static inline int roundup_nbits(size_t* out, const size_t in, const size_t nbits)
 {
     if (!out || nbits == 0)
     {
         return -EINVAL;
     }
-    if (in > UINT64_MAX - (nbits - 1))
+    if (in > SIZE_MAX - (nbits - 1))
     {
         return -EOVERFLOW;
     }
@@ -43,7 +43,7 @@ static inline int roundup_nbits(uint64_t* out, const uint64_t in, const uint64_t
     return 0;
 }
 
-static inline int rounddown_nbits(uint64_t* out, const uint64_t in, const uint64_t nbits)
+static inline int rounddown_nbits(size_t* out, const size_t in, const size_t nbits)
 {
     if (!out || nbits == 0)
     {
@@ -53,13 +53,13 @@ static inline int rounddown_nbits(uint64_t* out, const uint64_t in, const uint64
     return 0;
 }
 
-static inline int roundup_nbits_pow2(uint64_t* out, const uint64_t in, const uint64_t nbits)
+static inline int roundup_nbits_pow2(size_t* out, const size_t in, const size_t nbits)
 {
     if (!out || nbits == 0 || (nbits & (nbits - 1)) != 0)
     {
         return -EINVAL;
     }
-    if (in > UINT64_MAX - (nbits - 1))
+    if (in > SIZE_MAX - (nbits - 1))
     {
         return -EOVERFLOW;
     }
@@ -67,7 +67,7 @@ static inline int roundup_nbits_pow2(uint64_t* out, const uint64_t in, const uin
     return 0;
 }
 
-static inline int rounddown_nbits_pow2(uint64_t* out, const uint64_t in, const uint64_t nbits)
+static inline int rounddown_nbits_pow2(size_t* out, const size_t in, const size_t nbits)
 {
     if (!out || nbits == 0 || (nbits & (nbits - 1)) != 0)
     {
@@ -77,13 +77,13 @@ static inline int rounddown_nbits_pow2(uint64_t* out, const uint64_t in, const u
     return 0;
 }
 
-static inline int roundnearest_nbits_pow2(uint64_t* out, const uint64_t in, const uint64_t nbits)
+static inline int roundnearest_nbits_pow2(size_t* out, const size_t in, const size_t nbits)
 {
     if (!out || nbits == 0 || (nbits & (nbits - 1)) != 0)
     {
         return -EINVAL;
     }
-    if (in > UINT64_MAX - (nbits >> 1))
+    if (in > SIZE_MAX - (nbits >> 1))
     {
         return -EOVERFLOW;
     }
@@ -91,13 +91,13 @@ static inline int roundnearest_nbits_pow2(uint64_t* out, const uint64_t in, cons
     return 0;
 }
 
-static inline int roundnearest_nbits(uint64_t* out, const uint64_t in, const uint64_t nbits)
+static inline int roundnearest_nbits(size_t* out, const size_t in, const size_t nbits)
 {
     if (!out || nbits == 0)
     {
         return -EINVAL;
     }
-    if (nbits > UINT64_MAX / 2 || in >= UINT64_MAX - (nbits / 2))
+    if (nbits > SIZE_MAX / 2 || in >= SIZE_MAX - (nbits / 2))
     {
         return -EOVERFLOW;
     }
